feat: 클래스 분리를 위한 ExceptionUtils 클래스 추가
