+++ README.md
@@ -39,6 +39,7 @@ baseball
     - RandomNumber
     
   - utils
+    - ExceptionUtils
     - InputUtils
     
   - view
